Add a pair of schema for FRUS taxonomy tags
- The .rnc file (rnc = RelaxNG compact syntax) is about the structure
of the document. oXygen uses this to flag things like
missing/wrong/out-of-order elements and attributes. For example, it
indicates that the entry for frus1977-80v01 has no tags (only 2
unofficial-tags-tbd).
- The .rnc file also embeds the taxonomy so oXygen autosuggests them
and shows hints about their meaning. This also flags any bad IDs used
for tags.
- The .sch file (sch = Schematron) has a single rule, flagging any
volumes with fewer than 4 tags. We can adjust this #, but this seemed
like a good starting point.